apitest

+++ src/main/java/com/board/board/controller/UserController.java
@@ -19,7 +19,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User getUser(@PathVariable Integer id){
+    public User getUser(@PathVariable Long id){
         return userService.getUser(id);
     }
